Migrate useComponentSize to TypeScript
